removing some warning related to twilio not having the credentials
