+++ test/integration/01-setup/01-funds-bitcoind-lnds.spec.ts
@@ -13,6 +13,7 @@ import {
 } from "test/helpers"
 
 jest.mock("src/realtimePrice", () => require("test/mocks/realtimePrice"))
+jest.mock("src/phone-provider", () => require("test/mocks/phone-provider"))
 
 const defaultWallet = ""
