+++ test/integration/02-user-wallet/02-fees-onchain.spec.ts
@@ -1,6 +1,7 @@
 import { bitcoindClient, getUserWallet } from "test/helpers"
 
 jest.mock("src/realtimePrice", () => require("test/mocks/realtimePrice"))
+jest.mock("src/phone-provider", () => require("test/mocks/phone-provider"))
 
 let userWallet0, userWallet1
