+++ test/integration/02-user-wallet/02-receive-lightning.spec.ts
@@ -3,6 +3,7 @@ import { getHash } from "src/utils"
 import { checkIsBalanced, getUserWallet, lndOutside1, pay } from "test/helpers"
 
 jest.mock("src/realtimePrice", () => require("test/mocks/realtimePrice"))
+jest.mock("src/phone-provider", () => require("test/mocks/phone-provider"))
 
 let userWallet1
 let initBalance1
