+++ test/integration/02-user-wallet/02-receive-rewards.spec.ts
@@ -3,6 +3,7 @@ import { OnboardingEarn } from "src/types"
 import { checkIsBalanced, getUserWallet } from "test/helpers"
 
 jest.mock("src/realtimePrice", () => require("test/mocks/realtimePrice"))
+jest.mock("src/phone-provider", () => require("test/mocks/phone-provider"))
 
 let userWallet1
